GAP Peripheral -> GAP Central
Lots of changes here, but it mostly boils down to swapping the BLE functionality for the assistance server and wearable device. The assistance request service still runs on the wearable, but the wearable is now the central. The wearable will also have to be the central for the door lock as well, so this change means that two different modes of connection won't have to be managed.

The board will also now automatically start scanning when the assistance button has been pressed, and terminate the connection when the assistance request has been acknowledged.